fix: remove needless tags in structs, add some test files and fix related interface
database design : https://dpk0kv.yuque.com/qlgg9k/bgod7e/bglktg
api design: https://www.apifox.cn/apidoc/shared-d990cb23-7304-403c-9712-41828db6e674/api-33100973

+++ services/category.go
@@ -19,15 +19,15 @@ func (s *category) Create(c *model.Category) error {
 	return s.GetDefaultGormDB().Model(model.Category{}).Create(c).Error
 }
 
-func (s *category) ChangeCategoryStatus(id uint, status uint8) error {
+func (s *category) ChangeCategoryStatus(id uint8, status uint8) error {
 	return s.GetDefaultGormDB().Model(model.Category{}).Where("id = ?", id).Update("status", status).Error
 }
 
-func (s *category) ChangeCategoryName(id uint, name string) error {
+func (s *category) ChangeCategoryName(id uint8, name string) error {
 	return s.GetDefaultGormDB().Model(model.Category{}).Where("id = ?", id).Update("name", name).Error
 }
 
-func (s *category) ListChildrenByParentID(parentID uint) ([]*Category, error) {
+func (s *category) ListChildrenByParentID(parentID uint8) ([]*Category, error) {
 	var categories []*Category
 	result := s.GetDefaultGormDB().Where("parent_id = ?", parentID).Find(&categories)
 	err := result.Error
